audit tests moreve to auditor service spec
